renaming file to RawClient
