--- core/src/main/java/org/stellar/anchor/client/TempClient.java
@@ -1,54 +1,0 @@
-package org.stellar.anchor.client;
-
-import com.google.gson.annotations.SerializedName;
-import java.util.Set;
-import lombok.*;
-
-/**
- * Represents a temporary client configuration used exclusively for loading client details from
- * inline YAML configuration. This class is intended as a temporary workaround to facilitate the
- * deserialization of client configurations. Note: If you add any new fields to the CustodialClient
- * or NonCustodialClient classes, ensure that you also add those fields here to maintain
- * consistency.
- */
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class TempClient {
-  @NonNull String name;
-
-  ClientConfig.ClientType type;
-
-  Set<String> domains;
-
-  @SerializedName("signing_keys")
-  Set<String> signingKeys;
-
-  @SerializedName("callback_urls")
-  ClientConfig.CallbackUrls callbackUrls;
-
-  @SerializedName("allow_any_destination")
-  boolean allowAnyDestination = false;
-
-  @SerializedName("destination_accounts")
-  Set<String> destinationAccounts;
-
-  public CustodialClient toCustodialClient() {
-    return CustodialClient.builder()
-        .name(name)
-        .signingKeys(signingKeys)
-        .callbackUrls(callbackUrls)
-        .allowAnyDestination(allowAnyDestination)
-        .destinationAccounts(destinationAccounts)
-        .build();
-  }
-
-  public NonCustodialClient toNonCustodialClient() {
-    return NonCustodialClient.builder()
-        .name(name)
-        .domains(domains)
-        .callbackUrls(callbackUrls)
-        .build();
-  }
-}
